Handle COH event in a separate thread
Pull async config out into its own configuration file.
