+++ src/main/java/uk/gov/hmcts/reform/sscscorbackend/Application.java
@@ -1,9 +1,6 @@
 package uk.gov.hmcts.reform.sscscorbackend;
 
 import java.util.Properties;
-import java.util.concurrent.Executor;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.aop.interceptor.AsyncUncaughtExceptionHandler;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
@@ -13,17 +10,11 @@ import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.ComponentScan;
 import org.springframework.mail.javamail.JavaMailSender;
 import org.springframework.mail.javamail.JavaMailSenderImpl;
-import org.springframework.scheduling.annotation.AsyncConfigurer;
-import org.springframework.scheduling.annotation.EnableAsync;
-import org.springframework.scheduling.concurrent.ThreadPoolTaskExecutor;
 import org.springframework.web.client.RestTemplate;
 import uk.gov.hmcts.reform.authorisation.ServiceAuthorisationApi;
 import uk.gov.hmcts.reform.authorisation.generators.AuthTokenGenerator;
 import uk.gov.hmcts.reform.authorisation.generators.AuthTokenGeneratorFactory;
-import uk.gov.hmcts.reform.logging.exception.AlertLevel;
 import uk.gov.hmcts.reform.sscs.ccd.config.CcdRequestDetails;
-import uk.gov.hmcts.reform.sscscorbackend.exception.SscsCorBackendException;
-
 
 @SpringBootApplication
 @EnableCircuitBreaker
@@ -37,9 +28,7 @@ import uk.gov.hmcts.reform.sscscorbackend.exception.SscsCorBackendException;
         })
 @SuppressWarnings("HideUtilityClassConstructor") // Spring needs a constructor, its not a utility class
 @ComponentScan(basePackages = {"uk.gov.hmcts.reform"})
-@EnableAsync
-@Slf4j
-public class Application implements AsyncConfigurer {
+public class Application {
 
     public static void main(final String[] args) {
         SpringApplication.run(Application.class, args);
@@ -88,25 +77,5 @@ public class Application implements AsyncConfigurer {
         properties.put("mail.smtp.ssl.trust","*");
         javaMailSender.setJavaMailProperties(properties);
         return javaMailSender;
-    }
-
-    @Bean
-    @Override
-    public Executor getAsyncExecutor() {
-        ThreadPoolTaskExecutor executor = new ThreadPoolTaskExecutor();
-        executor.setCorePoolSize(2);
-        executor.setMaxPoolSize(4);
-        executor.setQueueCapacity(500);
-        executor.setThreadNamePrefix("COHEventHandler-");
-        executor.initialize();
-        return executor;
-    }
-
-    @Override
-    public AsyncUncaughtExceptionHandler getAsyncUncaughtExceptionHandler() {
-        return (ex, method, params) -> {
-            SscsCorBackendException exc = new SscsCorBackendException(AlertLevel.P3, ex);
-            log.error("Unhandled in COH thread exception", exc);
-        };
     }
 }
