+++ src/main/java/uk/gov/hmcts/reform/sscscorbackend/config/AsyncConfiguration.java
@@ -1,0 +1,37 @@
+package uk.gov.hmcts.reform.sscscorbackend.config;
+
+import java.util.concurrent.Executor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.aop.interceptor.AsyncUncaughtExceptionHandler;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.scheduling.annotation.AsyncConfigurer;
+import org.springframework.scheduling.annotation.EnableAsync;
+import org.springframework.scheduling.concurrent.ThreadPoolTaskExecutor;
+import uk.gov.hmcts.reform.logging.exception.AlertLevel;
+import uk.gov.hmcts.reform.sscscorbackend.exception.SscsCorBackendException;
+
+@Configuration
+@EnableAsync
+@Slf4j
+public class AsyncConfiguration implements AsyncConfigurer {
+    @Bean
+    @Override
+    public Executor getAsyncExecutor() {
+        ThreadPoolTaskExecutor executor = new ThreadPoolTaskExecutor();
+        executor.setCorePoolSize(2);
+        executor.setMaxPoolSize(4);
+        executor.setQueueCapacity(500);
+        executor.setThreadNamePrefix("COHEventHandler-");
+        executor.initialize();
+        return executor;
+    }
+
+    @Override
+    public AsyncUncaughtExceptionHandler getAsyncUncaughtExceptionHandler() {
+        return (ex, method, params) -> {
+            SscsCorBackendException exc = new SscsCorBackendException(AlertLevel.P3, ex);
+            log.error("Unhandled in COH thread exception", exc);
+        };
+    }
+}
